[chrome.fish] Add an alias for net unit tests.
I would really like to unify all the test functions into a single implementation, but there is no straightforward way to evaluate a dynamic binary path with correctly quoted arguments.

Not only would

    eval "./out/Release/$TEST_BINARY" $argv

be unsafe, but it actually completely fails in a common case for me, which is test arguments containing asterisks that are not to be interpreted by the shell.

See https://github.com/fish-shell/fish-shell/issues/154 for the relevant `fish` issues.