Why is it impossible write any secure, portable shell invocations? 😵‍💫
Apparently the `-S` flag to `env` is not portable, but it's fairly widely supported.

In *theory* the arguments to `bun` are not needed if it is guaranteed that the shebang will never be used in a file that `bun` can mistake for an argument other than the file to be executed. I am less enthused about about this footgun lying around as a code example, so I guess this is the best compromise for now.