Switch to `tailscale-app` cask and add a workaround for Tailscale completions.
I'm using `tailscale-app` instead of the previous combination (the Mac App Store version + `tailscale` CLI), because the latter borked my Tailscale connection when the binary and the app got ouf of sync. The `tailscale` CLI linked by `tailscale-app` is inside the `.app`, so it should prevent that issue.

See https://github.com/Homebrew/homebrew-cask/pull/223049 for why the workaround for completions is needed.