[git] Disable `ignoreRevsFile` for now.
This causes `git blame` to fail on repos without the specified file. There doesn't seem to be a way to say "use this file if it exists" despite the "ecosystem" having established a convention of using `.git-blame-ignore-revs` when it exists. The workarounds would probably involve wrapping `git`, so let's just avoid the feature for now.